pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mine"
version = "0.1.0"
description = "A modern Minesweeper game with a built-in solver."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "customtkinter>=5.2.0",
    "Pillow>=10.0.0",
]

[project.scripts]
mine = "mine.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mine"]
